Implement PackageIcon automatic packing
Just like our PackagePath implies Pack=true, we should likewise automatically pack a None/Content item if it was used as the PackageIcon property.

We support this for relative paths within the project as well as linked files (in case users use the same icon across multiple projects).

This means that in the most common case (a .jpg/.png) alognside the project file, setting the PackageIcon property to that file will Just Work.