Add warning when Authors == AssemblyName
The SDKs set this default in Microsoft.NET.DefaultAssemblyInfo.targets, but it's not really a very useful default (like the Description one).

So we flag this as a warning now.

We take over NG0102 since that wasn't active in any way (analyzers never get 4k chars of a property anyway).