Automatically pack dependencies satellite resources for private packages
Added a new `PackSatelliteDlls` property that can be set to `false` to avoid packing the satellite resource assemblies when the project is localized. This setting only applies when `PackBuildOutput=true`.

When a `PackageReference` has `PrivateAssets=all`, we assume all assets should be packed alongside the primary output (just as you would see them in the build output). Package references that include satellite resources will show up in the output directory, so it makes sense to also pack them in this case.