Never infer implicit reference for existing direct reference
We should never traverse and infer implicit transitive dependencies for packages that are already directly referenced by the project.

This fixes the scenario where the project has a direct dependency to a package and another direct dependency has a transitive dependency to the same one. By inferring the latter as an implicit dependency, we overwrite the pack attributes of the former.