Make NuGetizer targets buildTransitive to fix SponsorLink
For development-dependency packages, no assets flow across project
references, including build targets and package dependencies.
This is problematic for SponsorLink since it needs targets to properly
surface build properties to the analyzer that performs the checks.

By making NuGetizer build targets transitive, we fix the project to
project (P2P) scenario where one project installs NuGetizer and
referencing projects start failing the SponsorLink check because the
SL targets are missing. (since the package dependency on SL isn't
surfaced in that case).

By also including the SL targets in our package and conditionally
importing them if they haven't already been imported, we further
remove the changes that the missing dependency causes failures.

This might not be the optimal way to solve this, but it's the only
one that works consistently. This obviously causes P2P references
to a project consuming NuGetizer to *also* become nugetized
themselves. This should be a minor issue, however.