Make sure we can explicitly pack a privateassets/development dependency
As a package author, you might consume a development dependency for your own authoring, but also might want to take a dependency on it in your package. GitInfo is an example.

We want to make sure we honor explicit Pack=true even if a package has PrivateAssets=all (which has different use cases too).