Trim indent/newlines in text-based package metadata
This allows having newlines in the csproj while still compacting for packing.

The trimming involves:
* Removing blank lines at the beginning and ending of the value
* Determining indenting on the first line and using that to remove subsequent indenting in all following lines.
* Applying markdown rules to remove newlines: meaning only full paragraph separating newlines (blank) will be kept.
* Collapse multiple blank newlines into just one (when separating paragraphs).

This means you can indent nicely in the XML (as VS does by default) while still keeping a clean and compact representation in the package. For example:

```xml
    <Description>
      Create, edit and run multiple C# top-level programs in the same project,
      respecting per-file `#:package` references and `#:property` 😍
    </Description>
```
Would become a single line of text.