Expand image url as well as link url, use robust Markdig
Rather than the brittle regex-based matching we were doing, switch to proper markdown processing so we can better support arbitrary content within the readme without potentially breaking due to the flaky nature of the regex.

We were also not expanding the image url if it contained a link, so now we're more consistent.